Refactor GraqlShell (#2685)
* Change `GraqlShell` to use `RemoteGrakn`

* Remove websocket support from engine

* Enable inference flag

* Use config file

* Change port of SNB to use gRPC

* Extract ExternalEditor

* Extract HistoryFile

* Extract BatchLoader

* Extract `GraqlShellOptions`

* Extract License

* Propagate IOException better

* Extract option constants

* Propagate exceptions correctly

* Merge ErrorMessage classes

* Move all of shell into one package
