+++ grakn-bootup/src/main/java/ai/grakn/bootup/Graql.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.bootup;
 
-import ai.grakn.graql.GraqlShell;
+import ai.grakn.graql.shell.GraqlShell;
 import ai.grakn.migration.csv.CSVMigrator;
 import ai.grakn.migration.export.Main;
 import ai.grakn.migration.json.JsonMigrator;
@@ -26,6 +26,7 @@ import ai.grakn.migration.sql.SQLMigrator;
 import ai.grakn.migration.xml.XmlMigrator;
 import ai.grakn.util.GraknVersion;
 
+import java.io.IOException;
 import java.util.Arrays;
 
 /**
@@ -40,11 +41,11 @@ public class Graql {
      *
      * @param args
      */
-    public static void main(String[] args) {
+    public static void main(String[] args) throws IOException, InterruptedException {
         new Graql().run(args);
     }
 
-    public void run(String[] args) {
+    public void run(String[] args) throws IOException, InterruptedException {
         String context = args.length > 0 ? args[0] : "";
 
         switch (context) {
