+++ grakn-core/src/main/java/ai/grakn/util/ErrorMessage.java
@@ -209,7 +209,11 @@ public enum ErrorMessage {
     ATTRIBUTE_TYPE_NOT_SPECIFIED("No attribute type provided for compute query."),
     K_SMALLER_THAN_TWO("k can't be smaller than 2."),
     INSTANCE_DOES_NOT_EXIST("Instance does not exist in the subgraph."),
-    MAX_ITERATION_REACHED("Max iteration of [%s] reached.");
+    MAX_ITERATION_REACHED("Max iteration of [%s] reached."),
+
+    //--------------------------------------------- Shell Errors ---------------------------------------------------
+    COULD_NOT_CONNECT("Could not connect to Grakn. Have you run 'grakn server start'?"),
+    NO_VARIABLE_IN_QUERY("There was no variable specified in the query. Perhaps you forgot to escape `\\` a `$`?");
 
     private final String message;
